requirements.txt
python-dotenv==1.0.1
fastapi
uvicorn[standard]
pydantic
PyJWT
passlib[bcrypt]
libsql-experimental
slowapi
bcrypt
python-multipart
